Divide project's structure into 3 main modules
