+++ index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Tic-tac-toe</title>
-    <script src="./script.js" defer></script>
+    <script src="./src/main.js" type="module"></script>
     <link rel="stylesheet" href="./style.css">
     <link rel="icon" href="./images/tic-tac-toe.png">
 </head>
